# chess/docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "chess-app API",
        "description": "\u30aa\u30f3\u30e9\u30a4\u30f3\u5bfe\u6226\u30c1\u30a7\u30b9\u30a2\u30d7\u30ea\u306e\u30d0\u30c3\u30af\u30a8\u30f3\u30c9API\u3002\n\n`POST /auth/register` \u3067\u30a2\u30ab\u30a6\u30f3\u30c8\u3092\u4f5c\u6210\u3057\u3001\u8fd4\u5374\u3055\u308c\u305f\u30c8\u30fc\u30af\u30f3\u3092 \u53f3\u4e0a\u306e Authorize \u306b\u5165\u529b\u3057\u3066\u304f\u3060\u3055\u3044\u3002",
        "license": {
            "name": ""
        },
        "version": "0.1.0"
    },
    "paths": {
...
        "/users/{id}": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "\u30e6\u30fc\u30b6\u30fc\u306e\u516c\u958b\u30d7\u30ed\u30d5\u30a3\u30fc\u30eb\u3092\u53d6\u5f97\u3059\u308b",
                "description": "\u5bfe\u6226\u76f8\u624b\u306e\u8868\u793a\u540d\u53d6\u5f97\u306a\u3069\u306b\u4f7f\u3046\u3002\u30d1\u30b9\u30ef\u30fc\u30c9\u30cf\u30c3\u30b7\u30e5\u7b49\u306f\u542b\u307e\u306a\u3044\u516c\u958b\u60c5\u5831\u306e\u307f\u8fd4\u3059\u3002",
                "operationId": "get_user",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "\u30e6\u30fc\u30b6\u30fcID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u516c\u958b\u30e6\u30fc\u30b6\u30fc\u60c5\u5831",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserPublicResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "\u30e6\u30fc\u30b6\u30fc\u304c\u898b\u3064\u304b\u3089\u306a\u3044"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserPublicResponse": {
                "type": "object",
                "description": "GET /users/:id \u306e\u30ec\u30b9\u30dd\u30f3\u30b9(\u516c\u958b\u3057\u3066\u3088\u3044\u60c5\u5831\u306e\u307f)",
                "required": [
                    "id",
                    "username"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "username": {
                        "type": "string"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearer_auth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}